Create_Multiple_VMs/requirements.txt
msrestazure
azure-mgmt-resource
azure-mgmt-core
azure-identity
azure-mgmt-network
azure-core-exceptions